fix(ai-reports): cast units_sold to numeric before dividing for velocity
PostgreSQL was performing truncating integer division for the velocity
column (units_sold::integer / integer day-count) before ROUND(...,2)
ever ran, e.g. 7 units over 3 days returned 2 instead of 2.33. The
adjacent days_of_stock_left expression already cast the numerator
::numeric; apply the same cast to velocity.

Add a multi-day fixture to FastestProductsTest (7 units sold 3 days
ago, non-exact quotient) since both existing fixtures sell same-day
and never exercise a divisor other than 1, so they couldn't catch
integer-division truncation.

+++ backend/app/Http/Controllers/Api/V1/FastestProductsController.php
@@ -47,7 +47,7 @@ class FastestProductsController extends Controller
                 cat.name                                                                  AS category_name,
                 so.units_sold,
                 so.revenue,
-                ROUND(so.units_sold / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1), 2) AS velocity,
+                ROUND(so.units_sold::numeric / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1), 2) AS velocity,
                 COALESCE(st.total_stock, 0)                                               AS current_stock,
                 CASE
                     WHEN so.units_sold = 0 OR (so.units_sold::numeric / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1)) = 0 THEN NULL
